Renames getSubheadings to matchKeywordInSubheadings
